+++ R/dbSendQuery.R
@@ -10,11 +10,15 @@ NULL
 #' @rdname PrestoConnection-class
 #' @importFrom methods new
 #' @importMethodsFrom DBI dbSendQuery
+#' @param quiet If a progress bar should be shown for long queries (which run
+#'        for more than 2 seconds. Default to `getOption("rpresto.quiet")` which
+#'        if not set, defaults to `NA` which turns on the progress bar for
+#'        interactive queries.
 #' @export
 setMethod(
   "dbSendQuery", c("PrestoConnection", "character"),
-  function(conn, statement, ...) {
-    query <- PrestoQuery$new(conn, statement, ...)
+  function(conn, statement, ..., quiet = getOption("rpresto.quiet")) {
+    query <- PrestoQuery$new(conn, statement, quiet = quiet, ...)
     result <- query$execute()
     return(result)
   }